sysvar: add a variable to control import mode
