+++ pkg/sessionctx/variable/sysvar.go
@@ -3450,6 +3450,12 @@ var defaultSysVars = []*SysVar{
 			return nil
 		},
 	},
+	{Scope: vardef.ScopeGlobal, Name: vardef.TiDBEnableImportMode, Value: BoolToOnOff(vardef.DefTiDBEnableImportMode), Type: vardef.TypeBool,
+		SetGlobal: func(_ context.Context, s *SessionVars, val string) error {
+			switchImportMode(TiDBOptOn(val))
+			return nil
+		},
+	},
 }
 
 // GlobalSystemVariableInitialValue gets the default value for a system variable including ones that are dynamically set (e.g. based on the store)
